Animation and minor bug fixed in invRay CSG

+++ NM4PIG/Examples/Animations/createAnimation.sh
@@ -12,23 +12,25 @@
 #!/bin/bash
 
 # Move to executable directory
-cd ../../NM4PIG
+cd ../../
 
 # Generate 360 images
-for angle in $(seq 0 359); do
+for angle in $(seq 14 359); do
     angleNNN=$(printf "%03d" $angle)
-    echo "Creating image with angle ${angleNNN}"
-    dotnet run -- demo -s 4 -a $angle -pfm TemporaryImage.pfm -ldr "../Examples/DemoAnimation/image${angleNNN}.jpg" -o -rnd f &>/dev/null
-    rm -f TemporaryImage.pfm
+    if [ ! -f "Examples/Animations/image${angleNNN}.jpg" ]; then
+       echo "Creating image with angle ${angleNNN}"
+        dotnet run -- render --scene "Examples/Inputs/Other/ortho.txt" -pfm TemporaryImage.pfm -ldr Examples/Animations/image${angleNNN}.jpg -W 1080 -H 720 --declare-float angle:${angleNNN} --factor 1 &>/dev/null    
+        rm -f TemporaryImage.pfm 
+    fi
 done
 
 echo -e "\nI created all the images, now I proceed with creating the animation"
 
 # Move back
-cd ../Examples/DemoAnimation
+cd Examples/Animations/
 
 # Create animation
-ffmpeg -r 25 -f image2 -s 640x480 -i image%03d.jpg -vcodec libx264 -pix_fmt yuv420p coloured-Spheres.mp4 &>/dev/null
+ffmpeg -r 25 -f image2 -s 640x480 -i image%03d.jpg -vcodec libx264 -pix_fmt yuv420p worldAndWikishape.mp4 &>/dev/null
 
 # Clear useless files
 # rm -f image*.jpg
@@ -36,4 +38,4 @@ ffmpeg -r 25 -f image2 -s 640x480 -i image%03d.jpg -vcodec libx264 -pix_fmt yuv4
 echo -e "\nI created the animation in the file coloured-Spheres.mp4"
 
 # Create also gif file
-#ffmpeg -i coloured-Spheres.mp4 coloured-Spheres.gif
+ffmpeg -i worldAndWikishape.mp4 worldAndWikishape.gif
